% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/analyzeResult.R
\name{analyzeResult}
\alias{analyzeResult}
\title{Analyze performance of estimated time series.}
\usage{
analyzeResult(modelOutput, zero.val = 0.001, cens.level = 0.005,
  min.obs = 0.01, ErrLevs = seq(0.01, 1.5, 0.01), bins = seq(0, 1, 0.01),
  FlowStat = F)
}
\arguments{
\item{modelOutput}{A list of model results, in the format returned by
\code{\link{estDAR}} or \code{\link{estQPPQ}}.}

\item{zero.val}{(optional) The censoring value used to indicate a
censored streamflow value.  The default is \code{0.001}.}

\item{cens.level}{(optional) The value below which a streamflow value is
considered censored.  (Rounded down rather than up.)  The default is
\code{0.005}.}

\item{min.obs}{(optional) The minimum non-censored value of streamflow.
The default is \code{0.01}.}

\item{ErrLevs}{(optional) A vector of percent error values that should be
used to construct the cumulative distribution of absolute percent errors.
The default is \code{seq(0.01,1.5,0.01)}.}

\item{bins}{(optional) A vector of the bin limits that should be used to
characterize the the percent errors along the flow duration curve.  The
default is \code{seq(0,1,0.01)}.}

\item{FlowStat}{(optional) A logical indicating if the streamflow
statistics should be evaluated.  This will increase run time.  The default
is \code{FALSE}.}
}
\value{
A list of several elements:
\item{PerfMat}{A data frame of single-value performance metrics for each
target location.  See details.}
\item{CumDistErr}{\itemize{A list of two elements:
\item{Levels}{The input \code{ErrLevs}.}
\item{CumFreq}{A matrix summarizing the quantiles of the absolute percent
errors at each site.}}}
\item{ErrAlongFDC}{\itemize{A list of several elements:
\item{bins}{The input \code{bins}.}
\item{centers}{The plotting centers of each bin.}
\item{meanErr}{The mean percent error for each bin for each target location.}
\item{medianErr}{The median percent error for each bin for each target
location.}}}
}
\description{
The function \code{analyzeResult} conducts a first level analysis of
estimated time series.
}
\details{
The single-value performance metric calcuated include: The Nash-Sutcliffe
efficieny of daily streamflow, the Nash-Sutcliffe efficiency of the
logarithms of the daily streamflow, the root-mean-square error statistics
from \code{\link{rmse.like}} for natural and logarithm streamflows, the
average percent errors of the natural and logarithm streamflows and the
Pearson and Spearman correlations between observed and simulated streamflow.
}

