% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/applyDAR.R
\name{applyDAR}
\alias{applyDAR}
\title{Apply two version of Drainage Area Ratio}
\usage{
applyDAR(listofgages, startDate, endDate, limWY = 1, limDAR = c(-Inf, Inf),
  cv = list(fold = NA, dense = TRUE), flowStat = FALSE)
}
\arguments{
\item{listofgages}{A dataframe of sites to be used for analysis.  Sites are
character strings with at least 8 numeric characters.}

\item{startDate}{Specifies a particular date from which to begin retrieving
the streamflow data.  Format should be "YYYY-MM-DD".  If blank, all available
data will be retrieved.}

\item{endDate}{Specifies a particular date up to which to retrieve
the streamflow data.  Format should be "YYYY-MM-DD".  If blank, all available
data will be retrieved.}

\item{limWY}{(optional) The minimum number of water years required for a
site to be included. The default is \code{1}.}

\item{limDAR}{(optional) A numeric vector with two elements, indicating the
minimum and maximum drainage area ratio to be allowed for the applicaiton
of drainage area ratio. The default is \code{c(-Inf,Inf)}.}

\item{cv}{(optional) A list of two elements defining the parameters of
cross validation.
\itemize{
\item{fold}{An integer indicating how many folds should be used for
evaluation.  The default, \code{NA} or \code{1}, requests a leave-one-out
analysis.}
\item{dense}{A logical value indicating if the smaller or larger half of the
cross-validation fold should be used as the index.  The defulat, \code{TRUE},
 calls for the dense scenario using the larger half of the fold as
 the index network.}
}}

\item{flowStat}{(optional) A logical indicating if the streamflow
statistics should be evaluated.  This will increase run time.  The default
is \code{FALSE}.}
}
\value{
A list of two elements:
\item{cvList}{Corss-validated estiamted and observed streamflows.}
\item{anList}{basic analysis of estiamted streamflows.}
}
\description{
The function \code{applyDAR} conducts a cross-validated evaluation of
nearest-neighbor and map-correlation drainage area ratio.
}
\details{
Lorem ipsum...
}

