% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxP.R
\name{approxP}
\alias{approxP}
\title{Interpolate and Extrapolate Streamflows along Flow Duration Curve to Return
Probabilities}
\usage{
approxP(qts, fdc)
}
\arguments{
\item{qts}{A numerical vector containing a time series of streamflows.
Probabilities will not be estimated for non-positive values.}

\item{fdc}{A data frame of two numerical variables.  The first, \code{p},
contains streamflow probabilities corresponding to the streamflow values in
the second, \code{q}.  The probabilities can be either nonexceedance or
exceedance probabilities; the output will be of the same nature.
Non-positive values of \code{q} will be removed because of the conversion
to common logarithms.}
}
\value{
A list of three elements: \item{qts}{The input time series of
  streamflows.} \item{fdc}{The input flow duration curve, with addition of
  the common logarithm of streamflow (\code{ql}) and the standard nomral
  quantiles of the probabilities (\code{z}).} \item{targetP}{The resulting
  conversion of \code{qts} into streamflow probabilities by interpolating and
  extrapolating along the given flow duration curve.}
}
\description{
The function \code{approxP} interpolates and extrapolates a time
  series of streamflows along a given set of quantiles from a flow duration
  curve to produce streamflow probabilities.
}
\details{
The interpolation is accomplished by converting streamflows to the
  common logarithms of streamflow and converting the probabilites to standard
  normal quantlies.  The \code{approx} function is used to then linearly
  interpolate the standard normal quantiles against the common logarithms of
  streamflow.

  Streamflows in \code{qts} that are beyond the maximum value in \code{fdc$q}
  are extrapolated by extending the line between the two greatest quantiles
  of the flow duration curve.

  Similarly, streamflows in \code{qts} that are below the minimum value in
  \code{fdc$q} are extrapolated by extending the line between the two
  smallest quantiles of the flow duration curve.
}
\examples{
set.seed(1)
qts <- qnorm(runif(100), mean = 5, sd = 2)
fdc <- data.frame(
  p = seq(from = 0.01, to = 0.99, by = 0.01)
)
fdc$q <- qnorm(fdc$p, mean = 5, sd = 2)

result <- approxP(qts = qts, fdc = fdc)
}

