% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/best.models.R
\name{best.models}
\alias{best.models}
\title{Find top models for each flow regime.}
\usage{
best.models(top_n_list, unfilled_FDCs, expl, comp.wys, save.dir = NULL,
  nvmax = 6, n = 20, cens_level = 0.005, zero.val = 0.001,
  WY.lim = 10, regimes = list(lowflow = c("0.0002", "0.0005", "0.001",
  "0.002", "0.005", "0.01", "0.02", "0.05", "0.1"), medflow = c("0.2", "0.25",
  "0.3", "0.4", "0.5", "0.6", "0.7", "0.75", "0.8", "0.9"), highflow = c("0.95",
  "0.98", "0.99", "0.995", "0.998", "0.999", "0.9995", "0.9998")))
}
\arguments{
\item{top_n_list}{Output from \link{\code{compile.vars}}}

\item{unfilled_FDCs}{A matrix of the raw FDC quantiles for each site.
This is derived from the output of \link{\code{calcEmpFDCs}}.}

\item{expl}{A data frame of the potential explanatory variables,
derived from the output of \link{\code{getBasinChar}}.}

\item{comp.wys}{A vector indicating how many complete water years
were present for each site.  This is derived from the output of
\link{\code{calcEmpFDCs}}.}

\item{save.dir}{(optional) A directory to which results are to be saved.
The default behavior \code{save.dir=NULL} does not save results.}

\item{nvmax}{(optional)  The maximum number of variables that will be
considered in a regression.  The default is \code{6}.}

\item{n}{(optional) The maximum number of models subsetted.
The deafult is \code{20}.}

\item{cens_level}{(optional)  The value establishing left censoring.
The deafult is \code{0.005}.}

\item{zero.val}{(optional)  The value to which zeroes or negative quantiles
will be set to.  The default is \code{0.001}.}

\item{WY.lim}{(optional)  The minimum number of water years required to be
included in the formation of regional regressions.  Only reference-quality
sites with at least this many complete water years will be used.
The default is \code{10}.}

\item{regimes}{(optional) A list of three elements (\code{names(regimes) =
c('lowflow','medflow','highflow')}), where each element is a character
vector indicating the frequencies of each quantile. The default is \code{
list(lowflow = c("0.0002","0.0005","0.001","0.002","0.005","0.01","0.02",
"0.05","0.1"),medflow = c("0.2", "0.25", "0.3", "0.4", "0.5","0.6", "0.7",
 "0.75", "0.8", "0.9"),highflow = c("0.95", "0.98", "0.99", "0.995",
 "0.998", "0.999", "0.9995", "0.9998"))}.}
}
\value{
Top three models (by AIC and adjustes R2) for each regime.
}
\description{
Takes output from \link{\code{compile.vars}} and determines
the best average model for each regime.
}
\details{
This function outputs the top 3 models for each flow regime by both
adjusted R^2 and AIC (for a total of 6 models). It takes the average
adjusted R^2 and AIC across quantiles within a flow regime by each number
of variables, and whichever number of variables has the highest R^2 or
lowest AIC is the "best" type of model for the flow regime.
}

