% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/calcEmpFDCs.R
\name{calcEmpFDCs}
\alias{calcEmpFDCs}
\title{Calculate empirical FDCs for all sites}
\usage{
calcEmpFDCs(flow_data, quant.type = 9, cens = "unfilled",
  cens_level = 0.005, begWyr = 1900, endWyr = 2014, saveName = "",
  FDC.probs = c(2e-04, 5e-04, 0.001, 0.002, 0.005, 0.01, 0.02, 0.05, 0.1, 0.2,
  0.25, 0.3, 0.4, 0.5, 0.6, 0.7, 0.75, 0.8, 0.9, 0.95, 0.98, 0.99, 0.995, 0.998,
  0.999, 0.9995, 0.9998))
}
\arguments{
\item{flow_data}{A zoo object of streamflow data.}

\item{quant.type}{(optional) An integer, from 1 to 9, defining the
formula used to compute quantiles. The default is \code{9}.}

\item{cens}{(optional) A character string, either \code{'filled'} or
\code{'unfilled'}, specifying how censored values will be handled.
The default is \code{'unfilled'}.}

\item{cens_level}{(optional)  The number specifying the censoring level.
The default is \code{0.005}.}

\item{begWyr}{(optional) The lower limit of water years to consider.
The default is \code{1900}.}

\item{endWyr}{(optional) The upper limit of water years to consider.
The default is \code{2014}.}

\item{saveName}{(optional) A character string of a file name, without
extension, in which to save results.  An empty string, the default, requests
that nothing be saved.}

\item{FDC.probs}{(optional) THe probabilities of the quantiles to be
calculated.  The default is \code{c(0.0002,0.0005,0.001,0.002,0.005,0.01,
0.02,0.05,0.10,0.20,0.25,0.30,0.40,0.50,0.60,0.70,0.75,0.80,0.90,0.95,
0.98,0.99,0.995,0.998,0.999,0.9995,0.9998)}.}
}
\value{
A large list of many internal parameters.  The most salient are:
\item{empFDC}{A matrix of the empircal flow duration curve quantiles.}
\item{PORstats.compWyr}{The number of complete water
years available at each site.}
}
\description{
The function \code{nse} computes the Nash-Sutcliffe efficiency for the rows
or columns of a matrix.
}
\details{
This function computes flow duration curves and summary statistics for a
given time series (aka zoo object) of stream flows.
It outputs the FDCs directly, but can also write to external files
via \code{saveName}.

If \code{cens} is set to \code{'filled'}, \code{library(smwrQW)} is required.
(Install via instructions found here: https://github.com/USGS-R/smwrQW
as the package is not on CRAN.)
}

