% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/calcFlowStats.R
\name{calcFlowStats}
\alias{calcFlowStats}
\title{Calculate streamflow statistics.}
\usage{
calcFlowStats(Flow, Dates)
}
\arguments{
\item{Flow}{A vector of streamflow values.}

\item{Dates}{A corresponding vector of dates.}
}
\value{
A named vector of streamflow statistics.
}
\description{
The function \code{calcFlowStats} calculates a wide array of streamflow
statistics for a given time series.
}
\details{
Calculated statistics include: The daily median, the daily mean (FDSS 1),
the daily coefficient of variation, the annual coefficient of variation,
several streamflow percentiles (\code{percs=c(0.01,0.05,0.1,0.25,0.75,
0.9,0.95,0.99)}), the 90th percentile of annual maximums, the L-CV (FDSS 2),
the L-skew (FDSS 3), the L-kurtosis (FDSS 4), the lag-one autocorrelation (
FDSS 5), the amplitude and phase (FDSS 6 and 7) of the sinusoidal seasonal
 signal, and the 50th and 10th percentiles of the annual minimum seven-day
 average streamflow.
}

