% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_BCs.R
\name{clean_BCs}
\alias{clean_BCs}
\title{Clean out raw basin characteristics.}
\usage{
clean_BCs(BCs, BC.code6.remflg = F, max.BC.oneval.frac = 0.5, BC_suffix,
  destination = "")
}
\arguments{
\item{BCs}{A data frame of basin characteristics from
\code{\link{winnow_BCs}}.}

\item{BC.code6.remflg}{(optional)  A logical flag that indicates whether or
not variables derived from a non-calibrated hydrologic model will be removed.
  The default, \code{FALSE} requests that these be retained.}

\item{max.BC.oneval.frac}{(optional)  A constant between 0 and 1, such that
if at least that fraction of the BC values has a single value (usually zero)
then the BC will be removed.  The default is \code{0.5}.}

\item{BC_suffix}{(optional) A character string appending certain condition
onto saved file names.}

\item{destination}{(optional) A character string specfying a file name,
without an extension, to which to write intermediate results.  The default
is not to save results.}
}
\value{
A matrix of basin characteristics, with several removed.
}
\description{
The function \code{clean_BCs} cleans the raw basin characteristics by
removing unncessary or not-useful variables.
}
\details{
"Cleans" basin characteristics (BCs) by:
\itemize{
\item{1.}{\itemize{Removing BCs that are cannot be used because they are non-numeric
or, according to current thinking, should not be used because they:
\item{a)}{measure hydrologic modifications such as dams or land development}
\item{b)}{use NHD or TIGER roads which are spatially non-uniform
in their "resolution"}
\item{c)}{use measured streamflow (e.g, kriged streamflow statistics or
calibrated model output)}
\item{d)}{optionally, BCs based on non-calibrated hydrologic models}}}
\item{2.}{Removing BCs that have a large fraction, whose value is given by
\code{max.BC.oneval.frac},of values that are the same.}
}
}

