% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/comp.R
\name{comp}
\alias{comp}
\title{Compare the models with a specifc number of variables.}
\usage{
comp(flow, numvars, unfilled_FDCs, comp.wys, top_n_list, expl, WY.lim = 10,
  zero.val = 0.001, cens_level = 0.005, n = 20, plot.dir = NULL)
}
\arguments{
\item{flow}{A character string indicating the name of the flow regime.}

\item{numvars}{The maximum number of variables considered in
the regression.}

\item{unfilled_FDCs}{A matrix of the raw FDC quantiles for each site.
This is derived from the output of \link{\code{calcEmpFDCs}}.}

\item{comp.wys}{A vector indicating how many complete water years
were present for each site.  This is derived from the output of
\link{\code{calcEmpFDCs}}.}

\item{top_n_list}{Output from \link{\code{compile.vars}}}

\item{expl}{A data frame of the potential explanatory variables,
derived from the output of \link{\code{getBasinChar}}.}

\item{WY.lim}{(optional)  The minimum number of water years required to be
included in the formation of regional regressions.  Only reference-quality
sites with at least this many complete water years will be used.
The default is \code{10}.}

\item{zero.val}{(optional)  The value to which zeroes or negative quantiles
will be set to.  The default is \code{0.001}.}

\item{cens_level}{(optional)  The value establishing left censoring.
The deafult is \code{0.005}.}

\item{n}{(optional) The maximum number of models subsetted.
The deafult is \code{20}.}

\item{plot.dir}{A directory to which plots are to be saved.
The default behavior \code{save.dir=NULL} does not create plots.}
}
\value{
Unclear.  Used by \link{\code{best.models}}, among others.
}
\description{
A sub-function for single comparison of models in a flow regime.
}
\details{
Lorem ipsum...
}

