% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Cross-compare models.}
\usage{
compare(out, y, ranks, vars, names)
}
\arguments{
\item{out}{Unclear.}

\item{y}{Unclear.}

\item{ranks}{Unclear.}

\item{vars}{Unclear.}

\item{names}{Unclear.}
}
\value{
Unclear.
}
\description{
A sub-function used by \code{\link{compile.vars}}.
}
\details{
Lorem ipsum...
}

