% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/compYears.R
\name{compYears}
\alias{compYears}
\title{Function to determine complete water years}
\usage{
compYears(dates, data, year = c("water", "calendar", "climate"))
}
\arguments{
\item{dates}{A vector of \code{class(dates)='Date'}.}

\item{data}{Either a vector of the same length as \code{dates} or a
two-dimensional array with one dimension equal to the length of \code{dates}.}

\item{year}{A character string from \code{c('water','calendar','climate')}
indicating which type of year should be evaluated.  Water years run from
October 01 through September 30.  Calendar years run from January 01
through December 31.  Climate years run from April 01 through March 31.}
}
\value{
The function retruns a list of objects:
\item{logiY}{A logical data.frame of the same dimensions as \code{data}.
\code{TRUE} indicates that the element is a member of a complete wate year.}
\item{numY}{The number of complete years from each site.}
\item{sY}{A numeric vector naming the nominal year of
each element in \code{dates}.}
\item{year}{The same as the input, for validation.}
}
\description{
The function \code{compYears} looks through the data to determine which
years are full of real values.
}
\details{
Lorem ipsum...
}

