% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_target_FDC.R
\name{est_target_FDC}
\alias{est_target_FDC}
\title{Estimate a flow duration curve for an ungaged site.}
\usage{
est_target_FDC(best.mods, regSelect, target.empFDC.filled, target.regvar)
}
\arguments{
\item{best.mods}{A list of the best models for each flow regime.}

\item{regSelect}{Statistic used to select best model. The options
are \code{adjR2} or \code{AIC}.}

\item{target.empFDC.filled}{This is derived from the output of \code{\link{calcEmpFDCs}}.}

\item{target.regvar}{This is derived from the output of \code{\link{getBasinChar}}.}
}
\value{
Unclear.
}
\description{
A sub-function for construction of FDC for an ungaged site.
}
\details{
Lorem ipsum...
}

