% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDAR.R
\name{estDAR}
\alias{estDAR}
\title{Apply drianage-area ratio for a particular target and index set}
\usage{
estDAR(index.network, index.baschar, target.baschar, index.obs, target.obs,
  DAR.low = -Inf, DAR.high = Inf)
}
\arguments{
\item{index.network}{A numeric vector of the streamgage IDs to be used as
potential index sites.}

\item{index.baschar}{A data frame with the \code{LAT_GAGE_UTM},
\code{LNG_GAGE_UTM} and \code{DRAIN_SQKM} for each item of
\code{index.network$index}.}

\item{target.baschar}{Identical to \code{index.baschar} with respect to
\code{index.network$target}.}

\item{index.obs}{A zoo object of the observed streamflows for each
index gage.}

\item{target.obs}{A zoo object of the observed streamflows for each
target gage.}

\item{DAR.low}{The lower limit of the DAR that will be permitted for valid
application of DAR.  The default is \code{-Inf}.}

\item{DAR.high}{The upper limit of the DAR that will be permitted for valid
application of DAR.  The default is \code{Inf}.}
}
\value{
As output, the function returns a list of the same length as
\code{index.network$target}.  Each element represents a unique target
location.  Each element contains a data.frame consisting of
\item{date}{The date of the row.}
\item{est}{The estimated streamflow for each day.}
\item{ratio}{The ratio of the index and target drainage areas for each day.
This only changes when an alternative index is required.}
\item{indexflow}{The index streamflow for each day.}
\item{index}{The index ID used on each day.}
\item{metric}{The similarlity metric between the index and target.  This
depends on the metric using in the \code{index.network}; currently
functional for nearest-neighbor and map-correlation.}
\item{obs}{The observed streamflow at the target location on each day.
Used for validation.}
}
\description{
The function \code{estDAR} uses the ranked index network defined by
\code{index.network} to apply the drainage area ratio to predict streamflow
time series at each target location.
}
\details{
Lorem ipsum...
}
