% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estQPPQ.R
\name{estQPPQ}
\alias{estQPPQ}
\title{Apply QPPQ for a particular target and index set}
\usage{
estQPPQ(index.network, index.obs, index.empFDC, zero.val = NA, target.obs,
  target.empFDC, target.estFDC, pvals)
}
\arguments{
\item{index.network}{A numeric vector of the streamgage IDs to be used as
potential index sites.}

\item{index.obs}{A zoo object of the observed streamflows for each
index gage.}

\item{index.empFDC}{A matrix of the empirical quantiles for each site in
the \code{index.network$index}.}

\item{zero.val}{(optional) A value to replace zeros.  The deafult is
\code{NA}.}

\item{target.obs}{A zoo object of the observed streamflows for each
target gage.}

\item{target.empFDC}{A matrix of the empirical quantiles for each site in
the \code{index.network$target}.}

\item{target.estFDC}{A matrix of the estimated quantiles for each site in
the \code{index.network$target}.}

\item{pvals}{A vector of the p-values corresponding to the quantiles in
\code{index.empFDC}, \code{target.empFDC} and \code{target.estFDC}.}
}
\value{
As output, the function returns a list of the same length as
\code{index.network$target}.  Each element represents a unique target
location.  Each element contains a data.frame consisting of
\item{date}{The date of the row.}
\item{est}{The estimated streamflow for each day.}
\item{indexflow}{The index streamflow for each day.}
\item{index}{The index ID used on each day.}
\item{index.extrap}{A logical value indicating if the p-value of the index
station had to be extrapolated beyond the limits of the index flow duration
curve.}
\item{target.extrap}{A logical value indicating if the streamflow at the
target location had to be extrapolated beyond the limits of the target
flow duration curve.}
\item{metric}{The similarlity metric between the index and target.  This
depends on the metric using in the \code{index.network}; currently
functional for nearest-neighbor and map-correlation.}
\item{obs}{The observed streamflow at the target location on each day.
Used for validation.}
\item{obs.pval}{The observed p-value at the target location on each day.
Used for validation.}
}
\description{
The function \code{estQPPQ} uses the ranked index network defined by
\code{index.network} to apply the nonlinear spatial interpolation using
flow duration curves to predict streamflow time series at each target
location.
}
\details{
Lorem ipsum...
}
