% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasinChar.R
\name{getBasinChar}
\alias{getBasinChar}
\title{Collect and process basin characteristics for required basins.}
\usage{
getBasinChar(listofgages, basinChars = NULL, destination = "",
  BC.code6.remflg = F, max.BC.oneval.frac = 0.5, debug.flg = F)
}
\arguments{
\item{listofgages}{Dataframe of gage numbers (as a character vector)}

\item{basinChars}{(optional) A character vector that specifies certain
variables to retrieve.  The default is to retrieve all.}

\item{destination}{(optional) A character string specfying a file name,
without an extension, to which to write intermediate results.  The default
is not to save results.}

\item{BC.code6.remflg}{(optional)  A logical flag that indicates whether or
not variables derived from a non-calibrated hydrologic model will be removed.
  The default, \code{FALSE} requests that these be retained.}

\item{max.BC.oneval.frac}{(optional)  A constant between 0 and 1, such that
if at least that fraction of the BC values has a single value (usually zero)
then the BC will be removed.  The default is \code{0.5}.}

\item{debug.flg}{(optional) A logical indicating if verbose outputs are
required from variable transformation.  The default is \code{FALSE}.}
}
\value{
The function returns a list as output.  This list contains:
\item{DescBCs}{A data frame containing the station ID (\code{STAID}),
drainage area (\code{DRAIN_SQKM}), and latitude and longitude in Albers
projection (\code{LAT_GAGE_UTM} and \code{LNG_GAGE_UTM})
and decimal degrees (\code{LAT_GAGE} and \code{LNG_GAGE}) for the gage outlet
and basin centroid (\code{LAT_CENT} and \code{LoNG_CENT})
for the \code{listofgages}.}
\item{AllBCs}{A data frame containing all the variables in the database.}
\item{cleanBCs}{A data frame of basin characteristics screened to remove
categorical data or variables with a large proportion of
identical variables.}
\item{winnowedBCs}{A data frame of basin characteristics with correlations
removed and climates aggregated to seasonal signals.}
\item{transf}{\itemize{A list of the variable transformations:
  \item{transfBCs}{A data frame of basin characteristics with
  transformations for linearity applied.}
  \item{transf.info}{A data frame describing all transformations applied.}
  \item{transfBC.VIFs}{A vector of the variance inflation factors (VIFs)
  of each bsin characteristic.}
  \item{transfBC.corr}{A square matrix of the correlations between basin
  characteristics.}}
}
}
\description{
The function \code{getBasinChar} collects all of the basin characteristics
in GAGES-II for a specific \code{listofgages}.
}
\details{
Lorem ipsum...
}

