% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexMC.R
\name{indexMC}
\alias{indexMC}
\title{Rank the best index gages for each target site.}
\usage{
indexMC(index.gages, index.obs, index.baschar, target.gages, target.obs,
  target.baschar, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{index.gages}{A numeric vector of the streamgage IDs to be used as
potential index sites.}

\item{index.obs}{A zoo object of the observed streamflows for each
index gage.}

\item{index.baschar}{A data frame with the \code{LAT_GAGE_UTM},
\code{LNG_GAGE_UTM} and \code{DRAIN_SQKM} for each item of
\code{index.gages}.}

\item{target.gages}{A numeric vector of the streamgage IDs to be used as
target sites.}

\item{target.obs}{A zoo object of the observed streamflows for each
target gage.}

\item{target.baschar}{Identical to \code{index.baschar} with respect to
\code{target.gages}.}

\item{method}{A character string of either \code{"pearson"},
\code{"spearman"} and \code{"kendall"}, specifying which correlation
method to use.}
}
\value{
The function returns a list as output.  This list contains:
\item{Cor.Obs}{A matrix of the observed correlation between index and target
sites. The columns are in the order of \code{target.gages} and
the rows are in the order of \code{index.gages}.}
\item{Cor.Netw}{A matrix of the observed inter-correlation between index
sites. The columns are in the order of \code{target.gages} and
the rows are in the order of \code{index.gages}.}
\item{Cor.Est}{A matrix of the estimated correlation between index and target
sites. The columns are in the order of \code{target.gages} and
the rows are in the order of \code{index.gages}.}
\item{varPar}{A matrix of the variogram parameters (covariance parameters
and nugget) for the variogram built on each index gage.}
\item{Rank}{A matrix of the index IDs for each target gage.
The columns are in the order of \code{target.gages}.}
\item{index}{The input \code{index.gages}.}
\item{target}{The input \code{target.gages}.}
}
\description{
The function \code{indexMC} returns the ranking of potential index sites,
by map correlation, for each target location.
}
\details{
Lorem ipsum...
}
