% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/indexNN.R
\name{indexNN}
\alias{indexNN}
\title{Rank the best index gages for each target site.}
\usage{
indexNN(index.gages, index.baschar, target.gages, target.baschar)
}
\arguments{
\item{index.gages}{A numeric vector of the streamgage IDs to be used as
potential index sites.}

\item{index.baschar}{A data frame with the \code{LAT_GAGE_UTM},
\code{LNG_GAGE_UTM} and \code{DRAIN_SQKM} for each item of
\code{index.gages}.}

\item{target.gages}{A numeric vector of the streamgage IDs to be used as
target sites.}

\item{target.baschar}{Identical to \code{index.baschar} with respect to
\code{target.gages}.}
}
\value{
The function returns a list as output.  This list contains:
\item{Dist}{A matrix of the distances between index and
target sites.  The columns are in the order of \code{target.gages} and
the rows are in the order of \code{index.gages}.}
\item{DAR}{A matrix of the drainage area ratios between index and
target sites.  The columns are in the order of \code{target.gages} and
the rows are in the order of \code{index.gages}.}
\item{Rank}{A matrix of the index IDs for each target gage.
The columns are in the order of \code{target.gages}.}
\item{index}{The input \code{index.gages}.}
\item{target}{The input \code{target.gages}.}
}
\description{
The function \code{indexNN} returns the ranking of potential index sites,
by geographic proximity, for each target location.
}
\details{
Lorem ipsum...
}

