% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_log10.R
\name{mean_diff_log10}
\alias{mean_diff_log10}
\title{Compute the average differences of base-10 logs}
\usage{
mean_diff_log10(Obs, Est, MARGIN = 2)
}
\arguments{
\item{Obs}{A matrix of observations.}

\item{Est}{A matrix of estimates.  Must be of the same dimensions as
\code{Obs}.}

\item{MARGIN}{(optional) The margin across which to compute the
differences.  The default is to compute mean differences for each column
(\code{MARGIN=2}).}
}
\value{
A vector of average differences of base-10 logs.
}
\description{
The function \code{mean_diff_log10} computes the average differences of
the base-10 logs of the rows or columns of a matrix.
}
\details{
Lorem ipsum...
}
