% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/obs.sim.corr.R
\name{obs.sim.corr}
\alias{obs.sim.corr}
\title{Compute the correlation between observations and estimates}
\usage{
obs.sim.corr(Obs, Est, MARGIN = 2, methods = c("pearson", "spearman",
  "kendall"))
}
\arguments{
\item{Obs}{A matrix of observations.}

\item{Est}{A matrix of estimates.  Must be of the same dimensions as
\code{Obs}.}

\item{MARGIN}{(optional) The margin across which to compute the
efficiencies.  The default is to compute an efficiency for each column
(\code{MARGIN=2}).}

\item{methods}{A character string of character vector indication which
correlation technique to use.  Identical to the input passed to \code{cor}.}
}
\value{
A vector or matrix of correlations.
}
\description{
The function \code{obs.sim.corr} computes the correlation between
observations and estimates for the rows or columns of a matrix.
}
\details{
Lorem ipsum...
}

