% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.error.R
\name{percent.error}
\alias{percent.error}
\title{Compute the average percent errors}
\usage{
percent.error(Obs, Est, MARGIN = 2)
}
\arguments{
\item{Obs}{A matrix of observations.}

\item{Est}{A matrix of estimates.  Must be of the same dimensions as
\code{Obs}.}

\item{MARGIN}{(optional) The margin across which to compute the
efficiencies.  The default is to compute an efficiency for each column
(\code{MARGIN=2}).}
}
\value{
A vector of average percent errors.
}
\description{
The function \code{percent.error} computes the average percent errors for
the rows or columns of a matrix.
}
\details{
Lorem ipsum...
}
