% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCompileFlow.R
\name{readCompileFlow}
\alias{readCompileFlow}
\title{Function to read and compile streamflow data for analysis}
\usage{
readCompileFlow(listofgages, units = c("cfs", "cms"), dataset_name = "",
  startDate = "", endDate = "", checkmiss = FALSE, keepNAdays = TRUE)
}
\arguments{
\item{listofgages}{Dataframe of gage numbers (as a character vector)}

\item{units}{Specifies the units of the output.  Options include "cms" for
cubic meters per second and "cfs" for cubic feet per second}

\item{dataset_name}{Allows the user to specify a specific output location.
Output will be saved as an R workspace.  If no name is given, output data
will be returned but not saved.}

\item{startDate}{Specifies a particular date from which to begin retrieving
the streamflow data.  Format should be "YYYY-MM-DD".  If blank, all available
data will be retrieved.}

\item{endDate}{Specifies a particular date up to which to retrieve
the streamflow data.  Format should be "YYYY-MM-DD".  If blank, all available
data will be retrieved.}

\item{checkmiss}{Logical.  If \code{TRUE}, if any day is missing a
streamflow value for a given site, none of the data for that site will be
collected.  If \code{FALSE}, missing days will be allowed.}

\item{keepNAdays}{Logical.If \code{TRUE}, if there is a day of flow missing,
the date will be included with an \code{NA} streamflow value. If
\code{FALSE}, only dates with streamflow data will be collected.  This input
 is only valid if \code{checkmiss=FALSE}.}
}
\value{
The function returns a zoo object of observed streamflows.
}
\description{
The function \code{readCompileFlow} collects the daily streamflow information
available for a specified \code{listofgages}.
}
\details{
Lorem ipsum...
}
