% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimes.R
\name{regimes}
\alias{regimes}
\title{Determine Optimal Regimes Along FDC}
\usage{
regimes(unfilled_FDCs, BCs, zero_val = 0.001)
}
\arguments{
\item{unfilled_FDCs}{A matrix of the raw FDC quantiles for each site.
This is derived from the output of \code{\link{calcEmpFDCs}}.}

\item{BCs}{A data frame of basin characteristics for each site,
derived from the output of \code{\link{getBasinChar}}.}

\item{zero_val}{(optional)  The value to which zeroes or negative quantiles
will be set to.  The default is \code{0.001}.}
}
\value{
A matrix containing:
\item{rmse}{Root-mean-squared error.}
\item{rmsne}{Root-mean-square-normalized error.}
\item{cvrmse}{The coefficient of variation of the root-mean-squared error.}
\item{nrmse}{Normalized root-mean-squared error.}
}
\description{
The function \code{regimes} explores the length of the FDC and breaks it
into several regimes, if appropriate.
}
\details{
Lorem ipsum...
}
