% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/rmse.like.R
\name{rmse.like}
\alias{rmse.like}
\title{Compute the Root-Mean-Squared Error Statistics}
\usage{
\method{rmse}{like}(Obs, Est, MARGIN = 2)
}
\arguments{
\item{Obs}{A matrix of observations.}

\item{Est}{A matrix of estimates.  Must be of the same dimensions as
\code{Obs}.}

\item{MARGIN}{(optional) The margin across which to compute the
efficiencies.  The default is to compute an efficiency for each column
(\code{MARGIN=2}).}
}
\value{
A matrix containing:
\item{rmse}{Root-mean-squared error.}
\item{rmsne}{Root-mean-square-normalized error.}
\item{cvrmse}{The coefficient of variation of the root-mean-squared error.}
\item{nrmse}{Normalized root-mean-squared error.}
}
\description{
The function \code{rmse.like} computes a series of statistics related to
the root-mean-squared error.
}
\details{
Lorem ipsum...
}

