% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/tolist.R
\name{tolist}
\alias{tolist}
\title{List of EVs}
\usage{
tolist(names, vars, range, leaps_list)
}
\arguments{
\item{names}{Unclear.}

\item{vars}{Unclear.}

\item{range}{Unclear.}

\item{leaps_list}{Output from \code{\link{compute.leaps.for}}}
}
\value{
A list of subsets of EVs (y), their ranks (rank),
and the full set of EVs for each quantile
}
\description{
A sub-function used by \code{\link{compile.vars}}.
}
\details{
Lorem ipsum...
}

