% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/transform_BCs.R
\name{transform_BCs}
\alias{transform_BCs}
\title{Develop transformations of basin characteristics.}
\usage{
transform_BCs(BCs, debug.flg = F, BC_sfx = "", destination = "")
}
\arguments{
\item{BCs}{A data frame of basin characteristics from
\code{\link{winnow_BCs}}.}

\item{debug.flg}{(optional) A logical indicating if verbose outputs are
required from variable transformation.  The default is \code{FALSE}.}

\item{BC_sfx}{(optional) A character string appending certain condition
onto saved file names.}

\item{destination}{(optional) A character string specfying a file name,
without an extension, to which to write intermediate results.  The default
is not to save results.}
}
\value{
A list of four elements:
\item{finalBCs}{A data frame of the final, transformed basin
characteristics.}
\item{transfBC.info}{A data fram of information to recreate variable
transformations.  This includes the constant added before transformation,
the transformation used and the constant added after transformation.}
\item{finalBC.VIFs}{The variance inflation factors between basin
characteristics.}
\item{finalBC.corr}{The cross-correlation matrix of final, transformed
basin characteristics.}
}
\description{
The function \code{transform_BCs} develops transformations of basin
characteristics to induce linearity.
}
\details{
Receives the "winnowed" basin characteristics (BCs) produced by
\code{\link{winndow_BCs}} and then determines need for transformation of
BCs by computing some summary statistics and then skewness for original and
centered data (if needed) and different transformations (currently powers
of 1/10, 1/5, 1/3, 1/2, 2/3, 3/2, 2, 3, 5 and 10 plus log10 and 10^)
following addition of constants as needed to remove negative and
non-positive values as needed, depending on the transformation.
Then skewness information is used to choose the preferred transformation
as the one which gives the skewness nearest to zero.
Following selection of best transformation, post-transformation centering is
applied as needed.
}

