% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winnow_BCs.R
\name{winnow_BCs}
\alias{winnow_BCs}
\title{Winnow Basin Characteristics}
\usage{
winnow_BCs(BCs, BC.code6.remflg = F)
}
\arguments{
\item{BCs}{A data frame of basin characteristics from
\code{\link{clean_BCs.R}}.}

\item{BC.code6.remflg}{(optional) A logical value indicating if the code-6
variables, those derived from an uncalibrated hydrologic model,
should be removed.  The defautl is \code{FALSE}.}
}
\value{
A matrix of basin characteristics, with some removed.
}
\description{
The function \code{winnow_BCs} removes correlated basin characteristics
and aggregate monthly variables to seasonal variables.
}
\details{
Reads in the "clean" basin characteristics (BCs) produced by
\code{\link{clean_BCs.R}} and then "winnows" them by:
\itemize{
\item{1.}{"pre-removing" various highly correlated or unlikely to be useful
BCs, including "SITE", "STD", "CDL", forest NLCD variables other than total
forest (FORESTNLCD06).}
\item{2.}{\itemize{Add a few BCs are combinations of existing BCs, including:
\item{(a)}{July - Jan monthly temperatures (then all other monthly
temperatures are removed; annual average is also retained.)}
\item{(b)}{Phase of annual cycle of monthly water balance WB (WB5100...)
variables, then delete monthly WB variables but retain annual WB variable
WB5100_ANN_MM.}
\item{(c)}{Combining monthly PPT variables into quarterly
(DJF, MAM, JJA, SON)}
\item{(d)}{Combining woody and emergent NLCD fractions to a single total
wetland fraction; also creating a NLCD total wetland plus NLCD water
variable}
\item{(e)}{Adding ELEV_RANGE_M_BASIN variable as difference of
ELEV_MAX_M_BASIN and ELEV_MIN_M_BASIN}}}
}
}

