% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/compute.leaps.for.R
\name{compute.leaps.for}
\alias{compute.leaps.for}
\title{Determine Best Subsets Models}
\usage{
compute.leaps.for(unfilled_FDCs, filled_FDCs, comp.wys, expl, WY.lim = 10,
  zero.val = 0.001, cens_level = 0.005, max.VIF = 10, nvmax = 6,
  nb = 3, plunge.ahead = T)
}
\arguments{
\item{unfilled_FDCs}{A matrix of the raw FDC quantiles for each site.
This is derived from the output of \link{\code{calcEmpFDCs}}.}

\item{filled_FDCs}{A matrix of the censor-filled FDC quantiles
for each site.  This is derived from the output of \link{\code{calcEmpFDCs}}.}

\item{comp.wys}{A vector indicating how many complete water years
were present for each site.  This is derived from the output of
\link{\code{calcEmpFDCs}}.}

\item{expl}{A data frame of the potential explanatory variables,
derived from the output of \link{\code{getBasinChar}}.}

\item{WY.lim}{(optional)  The minimum number of water years required to be
included in the formation of regional regressions.  Only reference-quality
sites with at least this many complete water years will be used.
The default is \code{10}.}

\item{zero.val}{(optional)  The value to which zeroes or negative quantiles
will be set to.  The default is \code{0.001}.}

\item{cens_level}{(optional)  The value establishing left censoring.
The deafult is \code{0.005}.}

\item{max.VIF}{(optional)  The maximum variance inflation factor permitted
in a candidate model.  The default is \code{10}.}

\item{nvmax}{(optional)  The maximum number of variables that will be
considered in a regression.  The default is \code{6}.}

\item{nb}{(optional) The maximum number of models that will be considered
for each quantile.  The deafult is \code{3}.}

\item{plunge.ahead}{(optional) A logical vector indicating if zero values
should or should not be handled with the \code{zero.val}.
The default is \code{TRUE}.}
}
\value{
A list of lists, one for each quantile, with each of those
containing one array of models per variable level. The models are
automatically ranked by regsubsets' weighted least-sqaures method,
but the reported R^2 and BIC are calculated from \link{\code{censReg}}.
}
\description{
The function \code{compute.leaps.for} explores the wide
set of candidate best-subsets models for each quantile.
The function \code{compute.leaps.foreach} accomplishes the same thing using
a parallel \code{for} loop.
}
\details{
This function creates and ranks regression models. It runs regsubsets on
the filled FDCs, and then uses those models in \link{\code{censReg}} on the
unfilled FDCs. All of the FDCs are transformed using the common logarithm.
Any model with a variable that has a VIF higher than the user defined
limit \code{max.VIF} will be dropped. The models are automatically ranked
by regsubsets' weighted least-sqaures method, but the reported R^2 and BIC
are calculated from \link{\code{censReg}}.
}

